Fixing Code
Fix the clear checking method, where you it will print and hold the history of the clearing of the check basic on the account. Fix Account info, it will be for not be seen as an transaction, and you can see who or when someone check the account.

With the Depositor class, I just adding a new line to the toString. (Making it look better)

To see the account info, when need to type in there account number and there SSN to get access to there privacy information